Fix issues #870, #871, #872, #873
- #870: Add pre-commit hooks for lint/format (.husky/pre-commit)
- #871: Run cargo clippy --workspace -- -D warnings and fix all violations across contracts/crates
- #872: Remove confirmed unused brain-storm-shared dependency from credential_metadata/Cargo.toml
- #873: Standardize Prettier config and format all frontend/backend files

Closes #870
Closes #871
Closes #872
Closes #873

+++ contracts/certificate/src/lib.rs
@@ -352,7 +352,7 @@ impl CertificateContract {
             .persistent()
             .get(&key)
             .unwrap_or_else(|| soroban_sdk::vec![&env]);
-        ids.len() as u32
+        ids.len()
     }
 
     /// Check whether a certificate is transferable.
